+++ actions/category_admin.py
@@ -27,9 +27,9 @@ from wagtailorderable.modeladmin.mixins import OrderableMixin
 from aplans.context_vars import ctx_instance, ctx_request
 from aplans.utils import append_query_parameter
 
+from actions.blocks.mixins import ActionListPageBlockFormMixin
 from admin_site.utils import admin_req
 from admin_site.wagtail import (
-    ActionListPageBlockFormMixin,
     AplansAdminModelForm,
     AplansCreateView,
     AplansEditView,
